LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/font-relative-units-expected.txt


PASS em relative inline-size
PASS rem relative inline-size
PASS ex relative inline-size
FAIL rex relative inline-size assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
PASS ch relative inline-size
FAIL rch relative inline-size assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
PASS ic relative inline-size
FAIL ric relative inline-size assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
PASS lh relative inline-size
PASS rlh relative inline-size
PASS cap relative inline-size
FAIL rcap relative inline-size assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
